requirements-pi.txt
numpy==1.26.1
typing-extensions==4.8.0
sympy==1.12
Pillow==10.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
filelock==3.12.4
idna==3.4
Jinja2==2.8.1
MarkupSafe==0.11
mpmath==1.3.0
requests==2.31.0
networkx==3.2
urllib3==2.0.7
#
torch==2.0.1
torchvision==0.15.2
